+++ client/src/config/personAndGroupSearchTags.ts
@@ -126,6 +126,11 @@ export const relatedAccordions: IHalLinks = {
     tab: 'concepts',
     jsonSearchTerm: 'material',
   },
+  events: {
+    title: 'Related Events',
+    halLinkName: 'lux:agentRelatedEvents',
+    tab: 'events',
+  },
   concepts: {
     title: 'Related Concepts',
     halLinkName: 'lux:agentRelatedConcepts',
@@ -139,11 +144,6 @@ export const relatedAccordions: IHalLinks = {
   eventsCarriedOut: {
     title: 'Events Carried Out',
     halLinkName: 'lux:agentEventsCarriedOut',
-    tab: 'events',
-  },
-  eventsContainingObjects: {
-    title: 'Events Containing Objects Created By',
-    halLinkName: 'lux:agentEventsUsingProducedObjects',
     tab: 'events',
   },
   members: {